fixed collison problem. Lasers were only colliding with the first sprite it hit. Each laser had a single last collide time value, if we checked for a collision with another sprite, before that value expired then no collision would happen. Now each laser keeps a last collide time for every sprite it has collided with, this way it will be able to collide with more than one sprite at a time.

+++ src/spacegame/tilegame/MainMenu.java
@@ -328,7 +328,7 @@ public class MainMenu implements ActionListener, ChangeListener{
 		}else{
 			ShipSlider hitpointSlider;
 			hitpointSlider = new ShipSlider("hitpoint", Ship.HITPOINT_MIN, 
-					Ship.HITPOINT_MAX, parent.getMap().getPlayer().getHitpoints());
+					Ship.HITPOINT_MAX, (int) parent.getMap().getPlayer().getHitpoints());
 			//Create the label table
 			Hashtable labelTable = new Hashtable();
 			labelTable.put( new Integer( Ship.HITPOINT_MAX ), new JLabel(String.valueOf(Ship.HITPOINT_MAX)) );
